Add example of AKS attestation and secret provisioning
This example demonstrates how Gramine DCAP SGX attestation quote can be
generated and verified from within an Azure Kubernetes Service (AKS)
cluster. The example contains reference wrappers that enable Gramine to
use AKS components such as AESMD and DCAP quote provider libraries. This
example also contains a microservice deployment for the RA-TLS verifier
(secret provisioning service) that can be readily deployed to the AKS
cluster.

This example is based on the `ra-tls-secret-prov` example from the core
Gramine repo. To run the client and server apps from that example inside
the AKS cluster, we prepare two Docker images, one for the client and
one for the server. The client image must be graminized via GSC.

This example is Ubuntu-specific (tested version is Ubuntu 18.04).

Signed-off-by: Veena Saini <veena.saini@intel.com>