Workaround the packaging issue with v1.3
The package `gramine-dcap_1.3-1` has two bugs: a missing dependency on
CJSON and a mispackaging of Secret Prov library. This commit:
- adds CJSON installation explicitly, and
- switches to the bug-free package `gramine_1.3-1`.

Because we switch to the package `gramine_1.3-1` (that expects the SGX
driver under `/dev/sgx/enclave` instead of `/dev/sgx_enclave`), this
commit also changes the `docker run --device=...` mapping.

Signed-off-by: jkr0103 <jitender.kumar@intel.com>