[Curated-Apps] Use latest image tags
Redis, PyTorch, MySQL and MariaDB image tags are updated.

Redis 7.0.11+ is based on Debian 12 which is not supported in GSC yet,
hence using 7.0.10 which is based on Debian 11. Memcached 1.6+ uses the
eventfd syscall which is insecure in Gramine, hence keeping the current
tag 1.5-20.04_beta. TF-Serving is built locally and is latest based on
Ubuntu 20.04. OpenVINO Model Server and scikit-learn are already on
latest tags.

Signed-off-by: jkr0103 <jitender.kumar@intel.com>