[Integrations/azure] Re-enable proxy support in libcurl
For minimization purposes, libcurl is built with most features disabled.
One of the disabled features was proxy support. This made libcurl-based
dependencies, such as RA-TLS libs, unusable in proxy-restricted
environments (e.g. with a company-wide proxy). This commit re-enables
proxy support, which increases the size of `libcurl.a` by ~7%.

This commit is the same as commit "[curl] Re-enable proxy support in
libcurl" in the core Gramine repo.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>