[Integrations/intel] Add RA-TLS/SecretProv libs for ITA
This commit adds the sources, documentation and tests for
`libra_tls_verify_ita.so` and `libsecret_prov_verify_ita.so` libraries,
used for integration of Gramine with Intel Trust Authority (ITA,
formerly Project Amber).

The sources and the build system are largely based on the Microsoft
Azure Attestation (MAA) RA-TLS libs found in the same repository.
However, ITA turned out to have substantial differences from MAA, thus
we decided to have separate sources for MAA and ITA (in contrast to e.g.
creating ITA as a series of patches to MAA).

Co-authored-by: Gordon King <gordon.king@intel.com>
Co-authored-by: Scott Raynor <scott.raynor@intel.com>
Signed-off-by: Gordon King <gordon.king@intel.com>
Signed-off-by: Scott Raynor <scott.raynor@intel.com>
Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>